pyproject.toml
[project]
name = "smartshop-api"
version = "0.1.0"
description = "SmartShop - Price Comparison App"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "streamlit>=1.22.0",
    "requests>=2.26.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.5",
    "pymongo>=4.5.0",
    "motor>=3.3.1",
    "beanie>=1.21.0",
    "beautifulsoup4>=4.9.3",
    "fake-useragent>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "frontend"]
